intel-socfpga: adjust align size for partition in wic
Issue: LIN10-3704

The original align size is 4k, and the environment variable size is
also 4k. When run command "saveenv", u-boot witres environment variable
from the first section, length is 4k. In this way, the last 512 bytes
of environment variable are wrote into the first partition and cause
it is broken.
Therefore, adjust align size to 512k so that there is enough space to
save environment variable.

Signed-off-by: Meng Li <Meng.Li@windriver.com>