intel-socfpga: implement psci_system_off interface in u-boot
Issue: LINCD-5222

This patch is missed when upgrade u-boot, so get it back from
WRlinux LTS 10.19.
When execute "poweroff" command in Linux, smc instruction is
invoked and then enter el3 mode to execute system off processing.
But in current code, there is only a weak define of psci_system_off,
and vendor doesn't implement this interface to power off system.
So, only make a workaround so that straxi10 board is able to run
command "poweroff" smoothly in linux.

Signed-off-by: Meng Li <Meng.Li@windriver.com>