intel-socfpga: create MMC specific defconfig for stratix10 platform
Issue: LINUXEXEC-17782

For stratix10 platform, it only supports system booting from sd card.
But the default defconfig enables all the booting device drivers for
qspi flash, nand flash and sd card. In this way, there will be timeout
if corresponding device is not on stratix10 board. So, create a MMC specific
defconfig to instead of the default one so that booting smoothly without
timeout and save boot time.

Signed-off-by: Meng Li <Meng.Li@windriver.com>