intel-socfpga: create MMC specific defconfig for agilex platform
Issue: LINUXEXEC-17980

For agilex platform, it only supports system booting from sd card.
But the default defconfig enables all the booting device drivers for
qspi flash, nand flash and sd card. In this way, there will be timeout
if corresponding device is not on agilex board. So, create a MMC specific
defconfig to instead of the default one so that booting smoothly without
timeout and save boot time.

Signed-off-by: Meng Li <Meng.Li@windriver.com>