intel-socfpga: use specific atf image to generate bootloader image for each platform
Issue: LINUXEXEC-17980

Now, this bsp supports for both stratix10 and agilex platforms, so it
needs to build atf image separately for the 2 platforms, and integrate
the corresponding atf image to the bootloader of each platform. Based
on this ideal, rewrite the do_compile() to support for building atf image
of multiple platforms at one time. When build u-boot-socfpga, copy
them into u-boot source directory, and set correct atf filename of each
platform. In this way, the bootload image of each platform  will include
the corresponding atf image.

Signed-off-by: Meng Li <Meng.Li@windriver.com>